requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
python-dateutil>=2.9.0.post0
tzdata>=2024.1
